smart-investing-app/requirements.txt
numpy
pandas
tensorflow
yfinance
finnhub-python
google-cloud-language
newspaper3k
matplotlib
joblib
scikit-learn
dotenv